fix: update schema to be correct for PostgreSQL
Update the schema to use `BIGINT` over `INTEGER` for session starting
times, as PostgreSQL only uses 4 bytes for an `INTEGER` unlike SQLite.

+++ database/schema.sql
@@ -2,7 +2,7 @@ DROP TABLE IF EXISTS sessions;
 CREATE TABLE sessions (
 	id INTEGER PRIMARY KEY,
 	title TEXT,
-	start_time INTEGER,
+	start_time BIGINT,
 	remaining INTEGER
 );
